+++ changelog.md
@@ -1,5 +1,6 @@
 # v1.0.3 
-- Internal tweaks
+- Added additional info button to *Safe Mode* label in options menu
+- Minor tweaks
 
 # v1.0.2 
 - Fix issue with *Mock your 90%+ Fail* option fetching records